backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
geoalchemy2
pandas
numpy
scikit-learn
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
pytest
httpx
joblib
google-generativeai
python-dotenv